+++ src/debuginfo/debuginfo.h
@@ -31,8 +31,8 @@ class DebugInfo
 {
   public:
 
-    HRESULT ResolveBreakpoint(CORDB_ADDRESS modAddress, const Source &source, int sourceLine, PDB::GlobalFileIndex &globalFileIndex,
-                              std::vector<PDB::ResolvedBreakpoint> &resolvedPoints);
+    HRESULT ResolveBreakpoint(CORDB_ADDRESS modAddress, const Source &source, int32_t sourceLine, int32_t sourceColumn,
+                              PDB::GlobalFileIndex &globalFileIndex, std::vector<PDB::ResolvedBreakpoint> &resolvedPoints);
 
     using PDBInfoCallback = std::function<HRESULT(const PDBInfo &)>;
     HRESULT GetPDBInfo(CORDB_ADDRESS modAddress, const PDBInfoCallback &cb);